Refactored models to reduce duplicate code
